requirements.txt
numpy>=1.13.1
torch==1.3.1
scikit-learn
matplotlib
notebook
awkward
uproot_methods
uproot
progressbar2
